zynqmp: Add casts for printing intptr_t values

+++ sys/dev/clk/xilinx/zynqmp_reset.c
@@ -183,7 +183,7 @@ zynqmp_reset_assert(device_t dev, intptr_t id, bool reset)
 	if (id > ZYNQMP_RESET_MAX)
 		return (EINVAL);
 	sc = device_get_softc(dev);
-	device_printf(dev, "%s called for id = %ld, reset =%d\n", __func__, id, reset);
+	device_printf(dev, "%s called for id = %ld, reset =%d\n", __func__, (long)id, reset);
 	rv = ZYNQMP_FIRMWARE_RESET_ASSERT(sc->parent, id, reset);
 	return (rv);
 }
@@ -197,7 +197,7 @@ zynqmp_reset_is_asserted(device_t dev, intptr_t id, bool *reset)
 	if (id > ZYNQMP_RESET_MAX)
 		return (EINVAL);
 	sc = device_get_softc(dev);
-	device_printf(dev, "%s called for id = %ld\n", __func__, id);
+	device_printf(dev, "%s called for id = %ld\n", __func__, (long)id);
 	rv = ZYNQMP_FIRMWARE_RESET_GET_STATUS(sc->parent, id, reset);
 
 	return (rv);
